requirements.txt
Flask~=1.1.2
PyGithub~=1.54.1